backend/requirements.txt
fastapi
uvicorn
requests
python-dotenv
bcrypt
sqlalchemy
psycopg2-binary
alembic
pyjwt
passlib